split mcp server into manifest, prompts, schema-compile modules
